Rename clusters commands to databases
