+++ src/cli/databases_create.rs
@@ -1,4 +1,5 @@
-use crate::cli::cloud_json::JSONCloudClusterV3;
+use crate::cli::cloud_json::JSONCloudCreateClusterRequestV3;
+use crate::cli::util::find_project_id;
 use crate::client::CapellaRequest;
 use crate::state::State;
 use log::debug;
@@ -7,33 +8,37 @@ use std::sync::{Arc, Mutex};
 use tokio::time::Instant;
 
 use crate::cli::error::{
-    client_error_to_shell_error, deserialize_error, unexpected_status_code_error,
+    client_error_to_shell_error, no_active_project_error, serialize_error,
+    unexpected_status_code_error,
 };
-use crate::cli::util::NuValueMap;
 use nu_engine::CallExt;
 use nu_protocol::ast::Call;
 use nu_protocol::engine::{Command, EngineState, Stack};
 use nu_protocol::{Category, PipelineData, ShellError, Signature, SyntaxShape};
 
 #[derive(Clone)]
-pub struct ClustersGet {
+pub struct DatabasesCreate {
     state: Arc<Mutex<State>>,
 }
 
-impl ClustersGet {
+impl DatabasesCreate {
     pub fn new(state: Arc<Mutex<State>>) -> Self {
         Self { state }
     }
 }
 
-impl Command for ClustersGet {
+impl Command for DatabasesCreate {
     fn name(&self) -> &str {
-        "clusters get"
+        "databases create"
     }
 
     fn signature(&self) -> Signature {
-        Signature::build("clusters get")
-            .required("name", SyntaxShape::String, "the name of the cluster")
+        Signature::build("databases create")
+            .required(
+                "definition",
+                SyntaxShape::String,
+                "the definition of the cluster",
+            )
             .named(
                 "capella",
                 SyntaxShape::String,
@@ -44,7 +49,7 @@ impl Command for ClustersGet {
     }
 
     fn usage(&self) -> &str {
-        "Gets a cluster from the active Capella organization"
+        "Creates a new database on the active Capella organization"
     }
 
     fn run(
@@ -54,11 +59,11 @@ impl Command for ClustersGet {
         call: &Call,
         input: PipelineData,
     ) -> Result<PipelineData, ShellError> {
-        clusters_get(self.state.clone(), engine_state, stack, call, input)
+        clusters_create(self.state.clone(), engine_state, stack, call, input)
     }
 }
 
-fn clusters_get(
+fn clusters_create(
     state: Arc<Mutex<State>>,
     engine_state: &EngineState,
     stack: &mut Stack,
@@ -68,10 +73,10 @@ fn clusters_get(
     let span = call.head;
     let ctrl_c = engine_state.ctrlc.as_ref().unwrap().clone();
 
-    let name: String = call.req(engine_state, stack, 0)?;
+    let definition: String = call.req(engine_state, stack, 0)?;
     let capella = call.get_flag(engine_state, stack, "capella")?;
 
-    debug!("Running clusters get for {}", &name);
+    debug!("Running clusters create for {}", &definition);
 
     let guard = state.lock().unwrap();
     let control = if let Some(c) = capella {
@@ -80,43 +85,37 @@ fn clusters_get(
         guard.active_capella_org()
     }?;
     let client = control.client();
-
     let deadline = Instant::now().add(control.timeout());
-    let cluster = client
-        .find_cluster(name, deadline, ctrl_c.clone())
-        .map_err(|e| client_error_to_shell_error(e, span))?;
+
+    let project_name = match control.active_project() {
+        Some(p) => p,
+        None => {
+            return Err(no_active_project_error(span));
+        }
+    };
+    let project_id = find_project_id(ctrl_c.clone(), project_name, &client, deadline, span)?;
+
+    let mut json: JSONCloudCreateClusterRequestV3 = serde_json::from_str(definition.as_str())
+        .map_err(|e| serialize_error(e.to_string(), span))?;
+    json.set_project_id(project_id);
+
     let response = client
         .capella_request(
-            CapellaRequest::GetClusterV3 {
-                cluster_id: cluster.id(),
+            CapellaRequest::CreateClusterV3 {
+                payload: serde_json::to_string(&json)
+                    .map_err(|e| serialize_error(e.to_string(), span))?,
             },
-            deadline,
+            Instant::now().add(control.timeout()),
             ctrl_c,
         )
         .map_err(|e| client_error_to_shell_error(e, span))?;
-    if response.status() != 200 {
+    if response.status() != 202 {
         return Err(unexpected_status_code_error(
             response.status(),
             response.content(),
             span,
         ));
     };
-    let cluster: JSONCloudClusterV3 = serde_json::from_str(response.content())
-        .map_err(|e| deserialize_error(e.to_string(), span))?;
 
-    let mut collected = NuValueMap::default();
-    collected.add_string("name", cluster.name(), span);
-    collected.add_string("id", cluster.id(), span);
-    collected.add_string("status", cluster.status(), span);
-    collected.add_string(
-        "endpoint_srv",
-        cluster.endpoints_srv().unwrap_or_default(),
-        span,
-    );
-    collected.add_string("version", cluster.version_name(), span);
-    collected.add_string("tenant_id", cluster.tenant_id(), span);
-    collected.add_string("project_id", cluster.project_id(), span);
-    collected.add_string("cidr", cluster.place().cidr(), span);
-
-    Ok(collected.into_pipeline_data(span))
+    Ok(PipelineData::new_with_metadata(None, span))
 }
