+++ src/cli/mod.rs
@@ -18,15 +18,14 @@ mod cbenv_managed;
 mod cbenv_register;
 mod cbenv_unregister;
 mod cloud_json;
-mod clusters;
-mod clusters_create;
-mod clusters_drop;
-mod clusters_get;
-mod clusters_health;
 mod collections;
 mod collections_create;
 mod collections_drop;
 mod ctrlc_future;
+mod databases;
+mod databases_create;
+mod databases_drop;
+mod databases_get;
 mod doc;
 mod doc_get;
 mod doc_insert;
@@ -34,6 +33,7 @@ mod doc_remove;
 mod doc_replace;
 mod doc_upsert;
 mod fake_data;
+mod health;
 mod help;
 mod nodes;
 mod ping;
@@ -91,15 +91,14 @@ pub use buckets_update::BucketsUpdate;
 pub use cbenv_managed::CBEnvManaged;
 pub use cbenv_register::CbEnvRegister;
 pub use cbenv_unregister::CbEnvUnregister;
-pub use clusters::Clusters;
-pub use clusters_create::ClustersCreate;
-pub use clusters_drop::ClustersDrop;
-pub use clusters_get::ClustersGet;
-pub use clusters_health::HealthCheck;
 pub use collections::Collections;
 pub use collections_create::CollectionsCreate;
 pub use collections_drop::CollectionsDrop;
 pub use ctrlc_future::CtrlcFuture;
+pub use databases::Databases;
+pub use databases_create::DatabasesCreate;
+pub use databases_drop::DatabasesDrop;
+pub use databases_get::DatabasesGet;
 pub use doc::Doc;
 pub use doc_get::DocGet;
 pub use doc_insert::DocInsert;
@@ -107,6 +106,7 @@ pub use doc_remove::DocRemove;
 pub use doc_replace::DocReplace;
 pub use doc_upsert::DocUpsert;
 pub use fake_data::FakeData;
+pub use health::HealthCheck;
 pub use help::Help;
 pub use nodes::Nodes;
 pub use ping::Ping;
